split command file with commandDir
